+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 + Set a higher maximum length for username and password
++ Limited minimum password length to a maximum of 1024
 + Set a default value for `RESULTS_PER_PAGE` if not provided via `.env`
 
 ### Fixed